plat-rcar_gen4: Add HSCIF driver
Add HSCIF feature to support showing the secure log
through the UART-PHY terminal.
There are two cases of showing log:
1. CFG_SCIF = n : Showing log through the logging feature with
   RCAR_DEBUG_LOG = 1
2. CFG_SCIF = y : Showing secure log by serial

+++ core/include/console.h
@@ -1,6 +1,7 @@
 /* SPDX-License-Identifier: BSD-2-Clause */
 /*
  * Copyright (c) 2014, Linaro Limited
+ * Copyright (c) 2025, Renesas Electronics Corporation
  */
 
 #ifndef __CONSOLE_H
@@ -11,7 +12,7 @@
 
 
 void console_init(void);
-void console_putc(int ch);
+void console_putc(uint8_t ch);
 void console_flush(void);
 
 void plat_console_init(void);
